refactor: "Car 클래스 리팩토링" - 랜덤값 체크 후 move하도록 리팩토링
