+++ src/test/java/study/step3/CarTest.java
@@ -3,30 +3,18 @@ package study.step3;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
+import org.junit.jupiter.api.Test;
+import study.step3.domain.Car;
 
 public class CarTest {
 
-  @DisplayName("자동차의 전진 조건을 만족하지 못하여 전진하지 못한다.")
-  @ParameterizedTest
-  @ValueSource(ints = {0, 1, 2, 3})
-  public void move_Stopping_lessThan4(int input) {
+  @DisplayName("자동차는 랜덤하게 1칸 전진하거나 전진하지 않는다.")
+  @Test
+  public void move_forwardOrStop() {
     Car car = new Car();
 
-    car.move(input);
+    car.randomMove();
 
-    assertThat(car.getDistance()).isEqualTo(0);
-  }
-
-  @DisplayName("자동차의 전진 조건을 만족하여 전진한다.")
-  @ParameterizedTest
-  @ValueSource(ints = {4, 5, 6, 7, 8, 9, 10})
-  public void move_Moving_4OrMore(int input) {
-    Car car = new Car();
-
-    car.move(input);
-
-    assertThat(car.getDistance()).isEqualTo(1);
+    assertThat(car.getDistance()).isLessThanOrEqualTo(1);
   }
 }
